+++ lib/async/cable/railtie.rb
@@ -3,14 +3,14 @@
 # Released under the MIT License.
 # Copyright, 2024, by Samuel Williams.
 
-require_relative "middleware"
+require_relative "server"
 
 module Async
 	module Cable
-		# Rails integration that automatically inserts {Middleware} into the application middleware stack during initialization.
+		# Rails integration that configures Action Cable to use {Server}.
 		class Railtie < Rails::Railtie
-			initializer "async.cable.configure_rails_initialization" do |app|
-				app.middleware.use Async::Cable::Middleware
+			initializer "async.cable.configure_action_cable", before: "action_cable.set_configs" do |app|
+				app.config.action_cable.server = Async::Cable::Server
 			end
 		end
 	end